Split off central_unit coding - 'cherry-pick' from NoWarries!
